refactor: introduce attribute namespace for `AsTool`
