+++ src/ToolBox/Attribute/AsTool.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PhpLlm\LlmChain\ToolBox;
+namespace PhpLlm\LlmChain\ToolBox\Attribute;
 
 #[\Attribute(\Attribute::TARGET_CLASS | \Attribute::IS_REPEATABLE)]
 final readonly class AsTool
