+++ tests/ToolBox/Tool/ToolOptionalParam.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PhpLlm\LlmChain\Tests\ToolBox\Tool;
 
-use PhpLlm\LlmChain\ToolBox\AsTool;
+use PhpLlm\LlmChain\ToolBox\Attribute\AsTool;
 
 #[AsTool('tool_optional_param', 'A tool with one optional parameter', method: 'bar')]
 final class ToolOptionalParam
