+++ tests/ToolBox/Tool/ToolRequiredParams.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PhpLlm\LlmChain\Tests\ToolBox\Tool;
 
-use PhpLlm\LlmChain\ToolBox\AsTool;
+use PhpLlm\LlmChain\ToolBox\Attribute\AsTool;
 
 #[AsTool('tool_required_params', 'A tool with required parameters', method: 'bar')]
 final class ToolRequiredParams
